rename pkg from trace to blockdb
