+++ internal/blockdb/doc.go
@@ -1,0 +1,2 @@
+// Package blockdb saves chain and block data for inspection later to aid in debugging ibctest failures.
+package blockdb
