+++ internal/blockdb/migrate.go
@@ -1,4 +1,4 @@
-package trace
+package blockdb
 
 import (
 	"database/sql"
