+++ internal/blockdb/migrate_test.go
@@ -1,4 +1,4 @@
-package trace
+package blockdb
 
 import (
 	"testing"
