--- internal/trace/doc.go
@@ -1,2 +1,0 @@
-// Package trace saves data for inspection later to aid in debugging ibctest failures.
-package trace
